refactor options formatting
