+++ src/layout/FileUploadWithTag/EditWindowComponent.tsx
@@ -9,6 +9,7 @@ import { AttachmentActions } from 'src/features/attachments/attachmentSlice';
 import { FormLayoutActions } from 'src/features/layout/formLayoutSlice';
 import { useAppDispatch } from 'src/hooks/useAppDispatch';
 import { useAppSelector } from 'src/hooks/useAppSelector';
+import { useFormattedOptions } from 'src/hooks/useFormattedOptions';
 import { useLanguage } from 'src/hooks/useLanguage';
 import { AttachmentFileName } from 'src/layout/FileUpload/FileUploadTable/AttachmentFileName';
 import { FileTableButtons } from 'src/layout/FileUpload/FileUploadTable/FileTableButtons';
@@ -53,10 +54,7 @@ export function EditWindowComponent({
   const { id, baseComponentId, textResourceBindings, readOnly } = node.item;
   const { lang, langAsString } = useLanguage();
 
-  const translatedOptions = options?.map((option) => ({
-    ...option,
-    label: langAsString(option.label) as string,
-  }));
+  const formattedOptions = useFormattedOptions(options);
 
   const onDropdownDataChange = (attachmentId: string, value: string) => {
     if (value !== undefined) {
@@ -210,7 +208,7 @@ export function EditWindowComponent({
             <Select
               inputId={`attachment-tag-dropdown-${attachment.id}`}
               onChange={(value) => onDropdownDataChange(attachment.id, value)}
-              options={translatedOptions ?? []}
+              options={formattedOptions}
               disabled={saveIsDisabled}
               error={attachmentValidations.filter((i) => i.id === attachment.id).length > 0}
               label={langAsString('general.choose')}
